Add comprehensive SystemManager tests and fix implementation bugs
- Fix SystemManager.Tick to use pre-calculated indices for efficient stage lookup
- Fix CreateSystems array bounds bug (loop condition i <= tickStage)
- Change TickStage type from int to byte for consistency with ISystemFactory
- Add code analysis suppressions with justifications for CA1031
- Create Overlook.System.Tests project with 25 comprehensive unit tests
- Add test coverage for system creation, ticking, disposal, and error handling
- Add SystemManagerTests.cs with mock implementations for testing

+++ Packages/com.fullmetalbagel.overlook-system/SystemFactory.cs
@@ -19,7 +19,7 @@ public sealed partial class CsvFileSystemFactory : ISystemFactory
 
     public string SystemName { get; }
     public bool Enable { get; }
-    public int TickStage { get; }
+    public byte TickStage { get; }
     public int TickTimes { get; }
 
     public CsvFileSystemFactory(string row, ILogger<CsvFileSystemFactory> logger)
@@ -52,7 +52,7 @@ public sealed partial class CsvFileSystemFactory : ISystemFactory
 
         // Parse tick stage
         var tickStageStr = parts[3].Trim();
-        if (!int.TryParse(tickStageStr, out var tickStage))
+        if (!byte.TryParse(tickStageStr, out var tickStage))
         {
             LogInvalidTickStage(_logger, tickStageStr, SystemName);
             TickStage = 0; // Default to 0 if parsing fails
